The first thing we did was to import the user model with import {User} from '../../models/user'.
Then, notice a function load has been scaffolded for you to load the data you want, all you need to do is edit the path to the json data, which in this case is https://api.github.com/users/, and call your provider somewhere as GithubUsers.load().

We changed the line with this.http.get(''path/to/data.json), replace it so that now you have this.http.get('https://api.github.com/users'). We've also changed the property data from the scaffolded class to githubUsers.

Finally, we want results from Github to be converted into an array of users. We do this by casting the json result into an array of users with .map(res => <Array<User>>(res.json())), just below the this.http.get('https://api.github.com/users').